refactor: refactoring of the app component
