+++ src/components/App/index.tsx
@@ -1,40 +1,26 @@
-import React, {FC, useState} from 'react';
+import React, {FC} from 'react';
 import {View, FlatList} from 'react-native';
 import styles from './styles';
 import Navbar from '../Navbar';
 import FormAdd from '../FormAdd';
 import TodoItem from '../TodoItem';
+import useApp from './useApp';
 
 type TodoItemsType = {
   id: string;
   title: string;
 };
-type RenderItemType = {
+export type RenderItemType = {
   item: TodoItemsType;
   index: number;
 };
 
 const App: FC<any> = () => {
-  const [todoItems, setTodoItems] = useState<TodoItemsType[]>([]);
-
-  const addTodoItem = (title: string) => {
-    const newItem = {
-      id: Date.now().toString(),
-      title,
-    };
-    setTodoItems((oldTodoItems) => [...oldTodoItems, newItem]);
-  };
-
-  const deleteTodoItem = (id: string) => {
-    setTodoItems((oldTodoItems) =>
-      oldTodoItems.filter((item) => item.id !== id),
-    );
-  };
+  const {deleteTodoItem, addTodoItem, todoItems} = useApp();
 
   const renderItem = ({item, index}: RenderItemType) => (
     <TodoItem item={item} index={index} deleteTodoItem={deleteTodoItem} />
   );
-
   return (
     <>
       <Navbar title="P cool TODO app" />
